[Upd #151] Refactor to break circular dependency
