+++ src/models/modelsContext.ts
@@ -1,0 +1,17 @@
+export const FAILURE_MODE_CONTEXT = {
+};
+
+export const BEHAVIOR_CONTEXT = {
+};
+
+export const COMPONENT_CONTEXT = {
+};
+
+export const ABSTRACT_CONTEXT = {
+};
+
+export const EVENT_CONTEXT = {
+};
+
+export const FUNCTION_CONTEXT = {
+};
